requirements.txt
streamlit>=1.32.0
google-generativeai>=0.8.0
mcp>=1.0.0
nest_asyncio>=1.6.0
httpx>=0.27.0
